This reviews all the java.awt package.  Most of Mani's patches were fine to be accepted.
Several patches edited parts of the api that were package scoped.  This is a problem because it means any consumer of that API could potentially be binary incompatible.

The patch to Window.java was split into two - because most of it was fine, but some of it needed to be rejected.  Additionally there was a small mistake in MenuComponent.java that was fixed.